Zip file contains the whole code and story board for Hangman iOS game application. Also attached a video where you can see the execution in iphone 11 pro max simulator
In this iOS game application, the user selects a level of play, and he needs to randomly guess the words depending upon the number of alphabets available until he is out of his chances. Also makes different buzzer sounds depending upon selection of correct/wrong alphabet. The words are randomly picked depending upon the level of play from a Database.